Added BufferTyped class. Added BufferView Up/Download. Some unicode8 fixes. Added InitIndexBuffer and InverPermutation to GpuSort. Added Random Engine getter. Uint in sort shaders where applicable. Added more units tests for gpu sort and sph fluid.

+++ engine/shaders/src/compute/bitonicSortFloat/bigDispersePermutationFloat.comp.hlsl
@@ -1,9 +1,10 @@
 #include "computePushConstant.hlsli"
+#include "math.hlsli"
 
 
 
 #define BLOCK_SIZE 128
-RWStructuredBuffer<int> dataBuffer : register(u0);
+RWStructuredBuffer<float> dataBuffer : register(u0);
 RWStructuredBuffer<uint> permutationBuffer : register(u1);
 cbuffer Values : register(b2)
 {
@@ -19,7 +20,7 @@ void CompareAndSwap(uint i, uint j)
         return;
     if (dataBuffer[i] > dataBuffer[j])
     {
-        int tmp = dataBuffer[i];
+        float tmp = dataBuffer[i];
         dataBuffer[i] = dataBuffer[j];
         dataBuffer[j] = tmp;
